requirements.txt
streamlit
ultralytics
Pillow
numpy
opencv-python-headless
streamlit_paste_button